Removing two colums
Two columns is no bueno except on crazy large screens. So we will make
it single up to 1500px and see how that does.
A tag page doesn’t need to say “tag,” that seems redundant redundant.

I want users to be happy on the homepage if that is where their
journeys take them. They will get 600 words for their reward.